feat: 現状

+++ sdk/src/version.rs
@@ -4,30 +4,34 @@ pub const DEFAULT_API_VERSION: Version = Version {
     build: 14
 };
 
+#[repr(C)]
 #[cfg_attr(test, derive(Debug))]
 pub struct Version {
-    pub major: u8,
-    pub minor: u16,
-    pub build: u16
+    /// メジャーバージョン
+    pub major: u32,
+    /// マイナーバージョン
+    pub minor: u32,
+    /// ビルドナンバー
+    pub build: u32
 }
 
 impl Version {
     // 上位8ビットがメジャーバージョン
     #[inline]
-    pub fn get_major(version: u32) -> u8 {
-        (version >> 24) as u8
+    pub fn get_major(version: u32) -> u32 {
+        version >> 24
     }
 
     // 次の12ビットがマイナーバージョン
     #[inline]
-    pub fn get_minor(version: u32) -> u16 {
-        ((version & 0x00FFF000) >> 12) as u16
+    pub fn get_minor(version: u32) -> u32 {
+        (version & 0x00FFF000) >> 12
     }
 
     // 下位12ビットがビルドナンバー
     #[inline]
-    pub fn get_build(version: u32) -> u16 {
-        (version & 0x00000FFF) as u16
+    pub fn get_build(version: u32) -> u32 {
+        version & 0x00000FFF
     }
 }
 
@@ -45,6 +49,6 @@ impl From<u32> for Version {
 
 impl Into<u32> for Version {
     fn into(self) -> u32 {
-        ((self.major as u32) << 24)  | ((self.minor as u32) << 12) | self.build as u32
+        (self.major << 24)  | (self.minor << 12) | self.build
     }
 }
